Finish EconomyAPI

+++ plugins/EconomyAPI/resources/mysql.sql
@@ -1,12 +1,12 @@
 -- # !mysql
 -- # { economyapi
 -- #    { init
-CREATE TABLE IF NOT EXISTS money (name VARCHAR(30) NOT NULL, currency VARCHAR(30) NOT NULL, money FLOAT NOT NULL, transactionBlocked TINYINT(1) NOT NULL, PRIMARY KEY (name, currency))
+CREATE TABLE IF NOT EXISTS money (name VARCHAR(30) NOT NULL, currency VARCHAR(30) NOT NULL, money INT NOT NULL, transactionBlocked TINYINT(1) NOT NULL, PRIMARY KEY (name, currency))
 -- #    }
 -- #    { create
 -- #      :name string
 -- #      :currency string
--- #      :defaultMoney float
+-- #      :defaultMoney int
 INSERT IGNORE INTO money (name, currency, money, transactionBlocked) VALUES (:name, :currency, :defaultMoney, 0)
 -- #    }
 -- #    { get
@@ -17,7 +17,7 @@ SELECT * FROM money WHERE name = :name AND currency = :currency
 -- #    { update
 -- #      :name string
 -- #      :currency string
--- #      :money float
+-- #      :money int
 -- #      :transactionBlocked int
 UPDATE money SET transactionBlocked = :transactionBlocked, money = :money WHERE name = :name AND currency = :currency
 -- #    }
